requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
numpy>=1.24.0
faiss-cpu>=1.8.1
langchain>=0.0.208
pydantic>=1.10.7